Ensure we don't start the reporting thread until the first request comes in
Previously we tried to do it but some setups pre-loaded schema files on application start up and so the reporting thread started on the "master" process of preforking webservers like unicorn.

This ensures that although we instrument the schema right away (so it happens just once ideally), we only start the reporting thread at request time, which should be after the fork.